backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.30.0,<1
psycopg2-binary>=2.9.0,<3
